chore(api): move compatibility query builder to separate file
