+++ api/core/entities/provider_configuration.py
@@ -48,7 +48,7 @@ from models.provider import (
     TenantPreferredModelProvider,
 )
 from models.provider_ids import ModelProviderID
-from models.types import legacy_compatible_model_type_filter
+from models.utils.model_type_compat import legacy_compatible_model_type_filter
 
 logger = logging.getLogger(__name__)
