+++ api/core/ops/ops_trace_manager.py
@@ -44,7 +44,7 @@ from models.dataset import Dataset
 from models.model import App, AppModelConfig, Conversation, Message, MessageFile, TraceAppConfig
 from models.provider import Provider, ProviderCredential, ProviderModel, ProviderModelCredential, ProviderType
 from models.tools import ApiToolProvider, BuiltinToolProvider, MCPToolProvider, WorkflowToolProvider
-from models.types import legacy_compatible_model_type_filter
+from models.utils.model_type_compat import legacy_compatible_model_type_filter
 from models.workflow import WorkflowAppLog
 from tasks.ops_trace_task import process_trace_tasks
